+++ R/shared_class_functions.r
@@ -411,14 +411,25 @@ create <- function(x, ...) {
 #' }
 #'
 #' While key violations prevent insertion, re-insertion of existing records in
-#' an object element do not. This makes insertion equivalent to an `INSERT OR
-#' IGNORE` expression in SQL. In particular, it is somewhat like using this
+#' an object element does not. This makes insertion equivalent to an \code{INSERT OR
+#' IGNORE} expression in SQL. In particular, it is somewhat like using this
 #' expression in SQLite, since that implementation uses dynamic typing.
 #'
 #' If \code{vals} contains attributes not included in
-#' \code{\link{attrs_order}(x)}, \code{insert} throws an error. If a partial set
-#' of attributes are inserted, then data is only inserted into components of
-#' \code{x} whose required attributes are all inserted.
+#' \code{\link{attrs_order}(x)}, \code{insert} throws an error, since those
+#' attributes can't be inserted.
+#'
+#' If a partial set of attributes is inserted, and \code{all} is \code{FALSE},
+#' then data is only inserted into components of \code{x[relations]} whose
+#' required attributes are all present in \code{vals}. If \code{all} is
+#' \code{TRUE}, \code{insert} returns an error instead. This is useful when
+#' specifying \code{relations}: in that case, you often intend to insert
+#' into all of the specified elements, so not including all the required
+#' attributes is a mistake, and \code{all = TRUE} prevents it.
+#'
+#' If \code{all} is \code{TRUE}, \code{insert}
+#' throws an error in this case: This ensures you insert into all members of a
+#' specified value of \code{relations}.
 #'
 #' @param x a relational data object, into which to insert data, such as a
 #'   \code{\link{relation}} or \code{\link{database}} object.
@@ -426,12 +437,16 @@ create <- function(x, ...) {
 #' @param relations a character vector, containing names of elements of \code{x}
 #'   into which to insert data. By default, \code{insert} attempts to insert
 #'   data into every element.
+#' @param all a logical, indicating whether \code{vals} is required to contain
+#'   all attributes of all elements of \code{x[relations]}. By default, it is
+#'   not, and data is only inserted into elements of \code{x[relations]} whose
+#'   attributes are all present in \code{vals}.
 #' @param ... further arguments pass on to methods.
 #'
 #' @return An R object of the same class as \code{x}, containing the additional
 #'   new data.
 #' @export
-insert <- function(x, vals, relations = names(x), ...) {
+insert <- function(x, vals, relations = names(x), all = FALSE, ...) {
   UseMethod("insert")
 }